fix(tests): correct order type mismatch and toggle count in unit tests
WidgetConfig.kt:
- Change 'order to order.toLong()' in toMap() — test expects Long(0L)
  at line 82; storing as Int caused assertEquals(0L, Int(0)) to fail
  since Long.equals(Int) returns false on JVM

RegisterTriggerTest.kt:
- Change repeat(10) to repeat(9) — 10 is even so final value returns to
  initial (true); assertEquals(!true, true) always fails. With 9 (odd)
  toggles, final value is opposite of initial as the test comment states.

+++ app/src/main/java/com/orderMate/modals/WidgetConfig.kt
@@ -40,7 +40,7 @@ data class WidgetConfig(
         "isEnabled" to isEnabled,
         "isRequired" to isRequired,
         "showInFilter" to showInFilter,
-        "order" to order,
+        "order" to order.toLong(),
         "level" to level.name,
         "options" to options.map { it.toMap() }
     )
